general clean up of redundant variables.
Refactored REST API for reports and graphs.

Stubed out a report display user preferences hook that will set report 
params and then be overriden by params on the URL

Refactored the implementation of the dynamic where clauses in metrics 
to make site_id a constraint instead of hard coded into the query. This 
will allow reports to set site_id on the url.

expanded the REST API to pass in more params in anticipation of 
developing more rpeorts.