Fixing caching bug (#637)
executeQuery returns a truthy value if successful. In the postgresql case it is a string like "resource id='79' type='pgsql result'". I'm a bit puzzled why bools are tested against "=== true"  at all, and being a bool they can be tested just fine in an if, while the if also works for truthy and falsey values.